Feat: review page UI
