--- src/menu/review/index.jsx
@@ -1,10 +1,0 @@
-import React from 'react';
-import ReviewWrite from './component/reviewWrite';
-
-export default function ReviewPage() {
-    return (
-        <>
-            <ReviewWrite />
-        </>
-    );
-}
