--- src/menu/small_club/detail_page/_component/ReviewContent.jsx
@@ -1,3 +1,0 @@
-export default function ReviewContent() {
-    
-}
