+++ src/menu/small_club/detail_page/_component/ReviewTab.jsx
@@ -1,13 +1,23 @@
-import ReviewContent from "./ReviewContent";
+import { Link } from "react-router-dom";
 import ReviewStatics from "./ReviewStatics";
 import "./reviewTab.css";
+import ReviewBox from "./ReviewBox";
 
 export default function ReviewTab() {
     return (
         <div className="review_body">
-            <ReviewStatics />
+            <div className="statics_container">
+                <div className="statics_header">
+                    <h3>이런 점이 좋았어요!</h3>
+                    <Link className="review_write_btn" to="/menu/small_club/detail_page/review_write" style={{textDecoration: "none"}}>
+                        <img src="/buttons/write_review_icon.png" alt="write review"/>
+                        <p>리뷰쓰기</p>
+                    </Link>
+                </div>
+                <ReviewStatics />
+            </div>
             <div className="divider"></div>
-            <ReviewContent />
+            <ReviewBox />
         </div>
     )
 }
