--- src/menu/small_club/detail_page/_component/StaticBar.jsx
@@ -1,13 +1,0 @@
-import "./reviewTab.css";
-
-export default function StaticBar() {
-    return (
-        <>
-            <div className="statics_body">
-                <div className="static_bar"></div>
-                <div className="static_percentage"></div>
-                <p>😀 "분위기가 좋아요"</p>
-            </div>
-        </>
-    )
-}
