+++ src/menu/small_club/detail_page/_component/reviewStats.css
@@ -1,0 +1,33 @@
+.review-stats {
+    display: flex;
+    flex-direction: column;
+    gap: 8px;
+  }
+  
+  .bar-container {
+    width: 100%;
+    height: 35px;
+    background-color: #F2F2F2;
+    border-radius: 10px;
+    overflow: hidden;
+    /* position: relative; */
+  }
+  
+  .bar {
+    height: 100%;
+    background-color: #9EDBDB;
+    display: flex;
+    align-items: center;
+    padding: 0 8px;
+    border-radius: 12px;
+  }
+  
+  .text {
+    color: #000;
+    font-size: 15px;
+    font-weight: 700;
+    white-space: nowrap;
+    overflow: visible;
+    padding-left: 7px;
+    /* text-overflow: ellipsis; */
+  }
