+++ src/menu/small_club/detail_page/_component/reviewTab.css
@@ -1,16 +1,22 @@
 .divider {
-    width: 70%;
-    height: 2px;
-    background-color: #9C9C9C4D;
+    width: 140%;
+    border: 0.2px solid #9C9C9C4D;
+    margin: 40px 0px 40px 0px;
 }
 
 .statics_header {
     display: flex;
+    flex-direction: row;
     padding: 39px 0px 15px 0px;
 }
 
+.review_write_btn {
+    display: flex;
+    flex-direction: row;
+    padding-left: 150px;
+}
 
-.statics_header button {
+.statics_header p {
     background-color: transparent;
     border: none;
     cursor: pointer;
@@ -22,38 +28,9 @@
     align-items: center;
 }
 
-.statics_header img {
+.statics_container img {
     width: 21px;
     height: 21px;
     margin: 0px 0px 3px;
 }
 
-
-.static_bar {
-    position: relative;
-    /* width: 22%;
-    height: 3%; */
-    width: 250px;
-    height: 30px;
-    border-radius: 10px;
-    background-color: #F2F2F2;
-    margin-bottom: 12px;
-}
-
-.static_percentage {
-    position: absolute;
-    width: 200px;
-    height: 30px;
-    top: 507.5px;
-    border-radius: 10px;
-    background-color: #9EDBDB;
-}
-
-.statics_body p {
-    position: absolute;
-    top: 514px;
-    left: 220px;
-    font-size: 12px;
-    font-weight: 700;
-    line-height: 18px;
-}
