+++ src/menu/small_club/detail_page/detailPage.css
@@ -86,10 +86,12 @@
 .detail_tab button {
     border: none;
     background-color: transparent;
+    box-shadow: none;
     cursor: pointer;
     width: 10%;
     height: 105%;
     font-weight: 500;
+    text-align: center;
 }
 
 .detail_tab button.active{
